requirements.txt
psycopg2-binary
boto3
pandas
pyarrow
dotenv
supabase